Update the command name for the J-Link GDB server
Starting with version 6.30 of the J-Link tools visual front-ends have been added to the default JLinkGDBServer command - and the command line version has been replaced with the JLinkGDBServerCLExe (on OS X with the default installer this is symlinked as JLinkGDBServerCL in /usr/local/bin).

Updated the command to search for the following possible command names on the system path and retrun whichever is found first (or JLinkGDBServer if none is located):

* JLinkGDBServerCLExe
* JLinkGDBServerCL
* JLinkGDBServer

This should allow compatibility with the latest version and older versions on OS X and Linux. The windows commands do not appear to have been changed with the 6.30 release of the J-Link tools.