Add support for remote target ip in GDB for WSL2 environments
The reasoning behind this, is using wsl2 in windows has no access
to the usb port, so one has to use jlinkgdbserver on the host, however
due to some changes in the WSL2, localhost no longer has access to the
host's listening ports, but instead the "default gateway" address
has to be used for connections. Unfortunately that is changing on reboot.

Without killing the gdb server between debug sessions, weird side effects
can occur (especially on exceptions) -> normally it gets stuck, and
restarting the process is needed.

In this WSL2 scenario, one can use serverArgs -noLocalhostOnly to allow
remote access to the GDB server. And start the native windows executable
for JlinkGDBServerCL.exe in serverpath.

A better approach could be to introduce a wsl2 bool, which could
automatically get the default gw IP address and populate these two variables.