Add cortex-debug.pathsRelativeToHome config option
Setting it to true means that the armToolchainPath and openocdPath
config options have the system home prepended to them before use. This
is handy because VSCode doesn't have a cross-platform way to, in your
settings.json, point to a directory like ~/.comp2300.

More work is needed to make this config option apply to the other
cortex-debug settings we don't use.