SWO showOnStartup Fix
The current cortex-debug has an SWO parameter "showOnStartup" that is intended to flag for the "panel" view to be switched to the SWO output. However, this does not currently work. If the "panel" is switched to the SWO output immediately after it is created, an unchanging black view will appear. Manually switching the view to another and back will then show the SWO output. The reason for the unchanging back view is unknown.

This fix works-around the issue by adding a timer to the SWO console startup code. One second after the console is started the timer will then switch the "panel" to the SWO output view.

This work-around is a hack, but it was the only obvious fix.