Accept undefined stop reason
In some cases, the GDB client will not include a reason field in the
async stopped event. Most notably, the SEGGER JLink server program entry
stop event only does in some corner cases, and as a result, most JLink
launch sessions start in the reset code with the exception error widget.

Additionally, as the GDB class relied on the erronous "stopped" event to
set this.stopped, the setFunctionBreakpoints request running after the
"Initialized" event will deadlock waiting for a "generic-stopped" event
that already occurred.

Add a reason !== undefined guard for the stopped event in MI2 to fall
back to a simple "generic-stopped" event for this scenario. Mark the
debugger as "stopped" when the debugger is ready to avoid deadlocking
the configuration procedure without the invalid stop event.