Do not treat outOfBandRecords as responses
Checks incoming MI messages for resultRecords before treating them as
responses. In some cases, the server may issue outOfBandRecords with a
token, but no resultRecords. This can for instance happen when the
server wants to report progress on an event before it's finished.
JLink's target-download command does this, and without this extra check,
we'll go straight to `target halt` before the download is actually
finished.

JLink's target-download events look like this:

    {"token":8,"outOfBandRecord":[{"isStream":false,"type":"status","asyncClass":"download","output":[]}]}

Note the token (which matches the target-download command's token), and
the lack of a resultRecord. For my 250kB image, this event is issue 5
times before the resultRecord is issued.