Breakpoints: Send responses for all requested breakpoints (#490)
According to the DAP specification, the response to a Breakpoint
request should contain an entry for all requested breakpoints, including
those that won't be set.

Currently, the DAP server only returns entries for the breakpoints that
were successfully set, and returns an error if any of the break-insert
requests reported errors from the GDB server.

Change this to return unverified breakpoints for both erronous
break-inserts and ignored breakpoint requests. This makes the invalid
breakpoints show as greyed out in the UI, and removes the current error
toast for all debug setting. This is in line with the behavior of other
VS Code debuggers, like the NodeJS debugger.

Additionally fixes a bug in functionBreakpoints, where the debugger
returned the source parameter as a string, when it should be an object.